We must accept finite disappointment, but never lose infinite hope

+++ 0x0C-more_malloc_free/101-mul.c
@@ -31,7 +31,7 @@ int main(int argc, char *argv[])
 		}
 	}
 
-	result = atoi(argv[1]) * atoi(argv[2]);
+	result = atol(argv[1]) * atoi(argv[2]);
 	printf("%lu\n", result);
 	return (0);
 }
